features: integration #4. list pemeriksaan. #4a. list pemeriksaan detail. #2. list daftar ajuan diterima

+++ components/reguler/auditorHalal/layanan/DataPengajuan.vue
@@ -430,6 +430,10 @@ const deleteFactoryOrOutlet = async (type: string, el: any) => {
     url = '/reguler/pelaku-usaha/delete-legal'
     id = el.id_reg_legal
   }
+ else if (type === 'halal data') {
+    url = '/reguler/pelaku-usaha/delete-penyelia'
+    id = el.penyelia_id
+  }
   try {
     const response: any = await $api(url, {
       method: 'delete',
@@ -669,6 +673,7 @@ onMounted(async () => {
       :on-submit="() => triggerSaveModal('Penyelia Halal')"
       :on-add="() => triggerAddModal('Penyelia Halal')"
       :data="halalData"
+      :on-delete="(el: any) => deleteFactoryOrOutlet('halal data', el)"
       title="Penyelia Halal"
       with-add-button
     />
